+++ src/bin/dfufile-dump.rs
@@ -1,8 +1,8 @@
 //! Dumps the structure of the file given as argument
 
-use dfufile::DfuFile;
+use dfufile::{DfuFile, Error};
 
-fn main() -> Result<(), Box<dyn std::error::Error>> {
+fn main() -> Result<(), Error> {
     let path = std::env::args().nth(1).expect("No file given");
     let mut dfu_file = DfuFile::open(path)?;